Succesfully Completed Business Process
The major modification that were necessary were error handling routines. As there is simply too much IO in this routine for it to have a simple execution framework.

Errors : Succesfully Create In Google, Fail To Create In Database - Second Pass
This was fixed by checking from google to get the value, both in Group and Member Sections

Errors: Information Incomplete for Join. As the Group only had the email from the original group creation. So to remedy that we took what was a boolean and instead returned the HeadOption to get the row that is returned by the query and then used that information going forward, however no information will be retrieved unless it exists.